task/create a new file with name DogsController and moving the .cs file into it + create anothor file with name Dtos
