+++ Application/Dtos/DogDto.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Application.Dtos
+{
+	public class DogDto
+	{		public string Name { get; set; } = string.Empty;
+	}
+}
